refactor(utils): move joinPaths to filesystem module (#98)
Migrated the path joining logic from local static functions to the dedicated
filesystem utility. This centralization improves code cohesion and eliminates
duplication.

Key changes:
- Moved `joinPaths` implementation to `filesystem.hpp`.
- Updated `serveErrorFile` to use `utils::fs::joinPaths` for reliable
  resolution of error page paths against the server root.

Refs #80
