+++ src/http/Request.cpp
@@ -1,29 +1,10 @@
 #include "http/Request.hpp"
+#include "common/filesystem.hpp"
 #include "config/LocationBlock.hpp"
 #include "config/ServerBlock.hpp"
 #include <sstream>
 
 namespace http {
-
-namespace {
-std::string joinPaths(const std::string &p1, const std::string &p2) {
-    if (p1.empty() || p1 == "/") {
-        if (p2.empty() || p2[0] != '/')
-            return "/" + p2;
-        return p2;
-    }
-    if (p2.empty())
-        return p1;
-    std::string p1_clean = p1;
-    if (p1_clean[p1_clean.length() - 1] == '/') {
-        p1_clean.resize(p1_clean.length() - 1);
-    }
-    std::string p2_clean = p2;
-    if (p2_clean[0] != '/')
-        p2_clean = "/" + p2_clean;
-    return p1_clean + p2_clean;
-}
-} // namespace
 
 RequestStartLine::RequestStartLine() : method(RequestStartLine::UNKNOWN), version("HTTP/1.1") {}
 
@@ -139,10 +120,10 @@ std::string Request::resolvePath() const {
         if (reqPath.length() >= locPath.length()) {
             suffix = reqPath.substr(locPath.length());
         }
-        return joinPaths(aliasPath, suffix);
+        return utils::joinPaths(aliasPath, suffix);
     }
     std::string rootPath = loc->root();
-    return joinPaths(rootPath, reqPath);
+    return utils::joinPaths(rootPath, reqPath);
 }
 
 } // namespace http
